+++ packages/headless/src/app/engine.ts
@@ -32,6 +32,7 @@ import {analyticsMiddleware} from './analytics-middleware.js';
 import {configuration} from './common-reducers.js';
 import type {EngineConfiguration} from './engine-configuration.js';
 import {instantlyCallableThunkActionMiddleware} from './instantly-callable-middleware.js';
+import {generateAnswerListener} from './listener-middleware/generate-answer-listener-middleware.js';
 import type {LoggerOptions} from './logger.js';
 import {logActionErrorMiddleware} from './logger-middlewares.js';
 import {
@@ -376,6 +377,7 @@ function createMiddleware<Reducers extends ReducersMapObject>(
   ].concat(
     answerApi.middleware,
     answerGenerationApi.middleware,
+    generateAnswerListener.middleware,
     options.middlewares || []
   );
 }